# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "assignment-5-customer-data",
    "version": "1.0.0",
    "description": "API for managing customers",
    "contact": {}
  },
  "servers": [],
  "paths": {
...
    "/customers/{customerId}/contacts": {
      "post": {
        "summary": "Create a new contact for a specific customer",
        "description": "Create a new contact and associate it with a specific customer using the customer's ID.",
        "operationId": "createContact",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to associate with the new contact.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cnt_phone": {
                    "type": "string",
                    "description": "The phone number of the contact",
                    "example": "+123456789"
                  },
                  "cnt_fax": {
                    "type": "string",
                    "description": "The fax number of the contact",
                    "example": "+987654321"
                  }
                },
                "required": ["cnt_phone", "cnt_fax"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new contact for the customer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success create contact"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields must be filled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all contacts for a specific customer",
        "description": "Retrieve all contact information associated with a specific customer.",
        "operationId": "getAllContact",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer whose contacts are being retrieved.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all contacts for the specified customer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contact"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contacts not found for the specified customer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customerId}/contacts/{contactId}": {
      "get": {
        "summary": "Get a contact by ID for a specific customer",
        "description": "Retrieve a specific contact by its ID for a specific customer.",
        "operationId": "getContactById",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to whom the contact belongs.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "contactId",
            "in": "path",
            "required": true,
            "description": "The ID of the contact to retrieve.",
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the contact by ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/contacts/{id}": {
      "put": {
        "summary": "Update an existing contact",
        "description": "Update contact details for a specific contact by its ID.",
        "operationId": "updateContact",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the contact to update.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cnt_phone": {
                    "type": "string",
                    "description": "The phone number of the contact",
                    "example": "+123456789"
                  },
                  "cnt_fax": {
                    "type": "string",
                    "description": "The fax number of the contact",
                    "example": "+987654321"
                  }
                },
                "required": ["cnt_phone", "cnt_fax"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the contact.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success update contact"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "previousDataContact": {
                          "$ref": "#/components/schemas/Contact"
                        },
                        "updatedContact": {
                          "type": "object",
                          "properties": {
                            "cnt_phone": {
                              "type": "string",
                              "example": "+123456789"
                            },
                            "cnt_fax": {
                              "type": "string",
                              "example": "+987654321"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields must be filled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a contact by ID",
        "description": "Delete a specific contact by its ID.",
        "operationId": "deleteContact",
        "tags": ["Contacts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the contact to delete.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the contact.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success delete contact"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Contact"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Contact not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contact not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "cs_id": {
            "type": "integer",
            "description": "Customer ID"
          },
          "cs_fullName": {
            "type": "string",
            "description": "Full name of the customer"
          },
          "cs_email": {
            "type": "string",
            "description": "Email address of the customer"
          },
          "cs_gender": {
            "type": "string",
            "enum": ["Male", "Female"],
            "description": "Gender of the customer"
          },
          "cs_birthDate": {
            "type": "string",
            "format": "date",
            "description": "Birth date of the customer"
          }
        }
      },
...
      "Contact": {
        "type": "object",
        "properties": {
          "contact_id": {
            "type": "integer",
            "description": "The unique ID of the contact.",
            "example": 10
          },
          "contact_name": {
            "type": "string",
            "description": "The name of the contact.",
            "example": "John Doe"
          },
          "contact_email": {
            "type": "string",
            "description": "The email address of the contact.",
            "example": "john.doe@example.com"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        }
      },
...
    }
  }
}